requirements.txt
python-dotenv==1.0.0
redis==4.5.5
